--- docs/source/utils.rst
@@ -1,8 +1,0 @@
-Utilities
-=========
-
-Network Helpers
-^^^^^^^^^^^^^^^
-
-.. automodule:: nupic.engine.common_networks
-   :members:
